requirements.txt
mediapipe
opencv-python
numpy
ultralytics
torch
torchvision
tqdm
pyyaml
scikit-learn
seaborn
